addcontent

+++ src/method/controller.rb
@@ -18,12 +18,18 @@ module Adcontroller
                 csv << hash.values
             end
             p questions
-            puts "Another Question?".green
-            puts "Enter if want to continue".yellow
-            puts "[N] if want to go back".red
-            input = gets.chomp.downcase
-            if input == "n"
+            prompt = TTY::Prompt.new(active_color: :red)
+                choices = [
+                    {name:'Add More Question?', value: 1},
+                    {name:'Go Back', value: 2},
+                ]
+                user_input = prompt.select("Select an action?", choices)
+                case user_input
+                when 1
+                    print "\e[2J\e[f"
+                when 2
                 the_user_wants_to_quit = true
+                print "\e[2J\e[f"
             end
         end
     end
@@ -46,13 +52,15 @@ module Adcontroller
     end
 
     def backupfile
+        file_path = "data/quiz/survey.csv"
+        destination_directory = "data/quizbackup/"
+        FileUtils.cp(file_path, destination_directory)
         print "Type ".yellow
         print "Name".red
         print " to Backup Survey File".yellow
         puts ""
         renamebackup = gets.chomp.downcase
-        File.rename("data/quiz/survey.csv", "data/quizbackup/" + renamebackup + ".csv")
-        CSV.open("data/quiz/survey.csv", "a")
+        File.rename("data/quizbackup/survey.csv", "data/quizbackup/" + renamebackup + ".csv")
         print "\e[2J\e[f"
         sleep(0.7)
         puts "Backup Completed".yellow
@@ -62,13 +70,15 @@ module Adcontroller
     end
 
     def backupanswerfile
+        file_path = "data/answer/userans.csv"
+        destination_directory = "data/ansbackup/"
+        FileUtils.cp(file_path, destination_directory)
         print "Type ".yellow
         print "Name".red
         print " to Backup Answer File".yellow
         puts ""
         answerbackup = gets.chomp.downcase
-        File.rename("data/answer/userans.csv", "data/ansbackup/" + answerbackup + ".csv")
-        CSV.open("data/answer/userans.csv", "a")
+        File.rename("data/ansbackup/userans.csv", "data/ansbackup/" + answerbackup + ".csv")
         print "\e[2J\e[f"
         sleep(0.7)
         puts "Backup Completed".yellow
@@ -78,26 +88,41 @@ module Adcontroller
     end
 
     def deletecsv
-    print "\e[2J\e[f"
-    sleep(0.2)
-    print "Enter ".yellow
-    print "survey.csv".red
-    print " to delete".yellow
-    puts ""
-    delete_file_name = gets.chomp.downcase
-    if File.exist?("data/quiz/" + delete_file_name)
-        File.delete("data/quiz/" + delete_file_name)
-        print "File".yellow
-        print " #{delete_file_name}".green
-        print " has deleted".yellow
-        puts ""
-        else
+        prompt = TTY::Prompt.new(active_color: :red)
         print "\e[2J\e[f"
-        sleep(0.2)
-        print "File".red
-        print " #{delete_file_name}".yellow
-        print " does not exist".red
-        puts ""
+        choices = [
+            {name:'Do you need to delete survey?', value: 1},
+            {name:'Go Back', value: 2},
+        ]
+        user_input = prompt.select("Select an action?", choices)
+        case user_input
+        when 1
+            print "\e[2J\e[f"
+            sleep(0.2)
+            print "Enter ".yellow
+            print "survey.csv".red
+            print " to delete".yellow
+            puts ""
+            delete_file_name = gets.chomp.downcase
+            if File.exist?("data/quiz/" + delete_file_name)
+                File.delete("data/quiz/" + delete_file_name)
+                print "File".yellow
+                print " #{delete_file_name}".green
+                print " has deleted".yellow
+                puts ""
+                else
+                print "\e[2J\e[f"
+                sleep(0.2)
+                print "File".red
+                print " #{delete_file_name}".yellow
+                print " does not exist".red
+                puts ""
+                puts "Press enter to go back menu.".green
+                gets
+                print "\e[2J\e[f"
+            end
+        when 2
+                print "\e[2J\e[f"
         end
     end
